+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 All notable changes to `laravel-actionable-model` will be documented in this file.
 
+## 1.1.0 - 2021-03-22
+
+- Now, you can eager load your actions. `User::with('actions')` 
+- Add a new method `actionsHandler` to your models to help you filter the actions.
+
 ## 1.0.0 - 2021-03-21
 
 - initial release